Changes: -Added support for the new system of Tables

+++ data/posts.go
@@ -1,7 +1,7 @@
 package model
 
 type Post struct {
-	ID       int
+	ID       int64
 	Username string
 	Title    string
 	Content  string
